+++ src/rdkafka_transport.h
@@ -60,7 +60,7 @@ int rd_kafka_transport_poll(rd_kafka_transport_t *rktrans, int tmout);
 #if WITH_SSL
 void rd_kafka_transport_ssl_ctx_term (rd_kafka_t *rk);
 int rd_kafka_transport_ssl_ctx_init (rd_kafka_t *rk,
-				     char *errstr, int errstr_size);
+				     char *errstr, size_t errstr_size);
 #endif
 void rd_kafka_transport_term (void);
 void rd_kafka_transport_init(void);
